history cell working
